ExchangeAlgebra

Description

Released under the OWL license

Package for Exchange Algebra defined by Hiroshi Deguchi.

Exchange Algebra is an algebraic description of bookkeeping systems. Details are below.

https://www.springer.com/gp/book/9784431209850

https://repository.kulib.kyoto-u.ac.jp/dspace/bitstream/2433/82987/1/0809-7.pdf

_Note_ : The current version 0.1.0.0 will be completely changed shortly, especially in the accounts settings section.

Documentation

module ExchangeAlgebra.Algebra

module ExchangeAlgebra.Algebra.Transfer

module ExchangeAlgebra.Write

module ExchangeAlgebra.Simulate