ExchangeAlgebra.Algebra.Transfer

Description

Released under the OWL license

Package for Exchange Algebra defined by Hiroshi Deguchi.

Exchange Algebra is an algebraic description of bookkeeping systems. Details are below.

https://www.springer.com/gp/book/9784431209850

https://repository.kulib.kyoto-u.ac.jp/dspace/bitstream/2433/82987/1/0809-7.pdf

_Note_ : The current version 0.1.0.0 will be completely changed shortly, especially in the accounts settings section.

Documentation

data TransTable n b where

Transfer transformation table

Constructors

NullTable :: (HatVal n, HatBaseClass b) => TransTable n b

TransTable

Fields

:: (HatVal n, HatBaseClass b)

=> { _size :: Size

, _before :: b

Base before transformation

, _transFunc :: n -> n

Value transformation function

, _after :: b

Base after transformation

, _left :: TransTable n b

, _right :: TransTable n b

} -> TransTable n b

isNullTable :: TransTable n b -> Bool

Tests whether the transfer table is empty.

Complexity: O(1)

transfer :: (HatVal n, HatBaseClass b) => Alg n b -> TransTable n b -> Alg n b

table :: (HatVal n, HatBaseClass b) => [(b, b, n -> n)] -> TransTable n b

make TransTable from list same as fromList >>> table $ Hat::- Hat:|% (id :: NN.Double - NN.Double) ++ Hat::- Hat:<(Cash) |% id [(Hat:Cash,Hat:<Building,<function),(Hat:Building,Hat:<Cash,<function)]

data TransTableParts b

A part of a transfer rule. Represents a source-to-target base pair in from :-> to form.

(.->) :: HatBaseClass b => b -> b -> TransTableParts b infixr 8

Transfer rule construction operator. from .-> to produces a TransTableParts.

Complexity: O(1)

(|%) :: (HatVal n, HatBaseClass b) => TransTableParts b -> (n -> n) -> [(b, b, n -> n)] infixr 7

Syntax to make list for makeList

>>> Hat:<(Yen,Cash):-> Hat:<(Yen,Building) |% (id :: NN.Double -> NN.Double) ++ Not:<(Yen,Building)  :-> Not:<(Yen, Cash)  |% id
[(Hat:<(Yen,Cash),Hat:<(Yen,Building),<function>),(Not:<(Yen,Building),Not:<(Yen,Cash),<function>)]

createTransfer :: (HatVal n, ExBaseClass b) => [(b, b, n -> n)] -> Alg n b -> Alg n b

Build an indexed fast transfer function from a list of transfer rules. More efficient than transfer when repeatedly applying the same TransTable.

Complexity: Build O(r log r) (r = number of rules); apply O(s) (s = number of entries)

incomeSummaryAccount :: (HatVal n, ExBaseClass b) => Alg n b -> Alg n b

Income Summary Account: compute net income for the current period.

netIncomeTransfer :: (HatVal n, ExBaseClass b) => Alg n b -> Alg n b

Net income transfer. Transfers NetIncome/NetLoss to RetainedEarnings.

Complexity: O(s) (s = total number of scalar entries)

grossProfitTransfer :: (HatVal n, ExBaseClass b) => Alg n b -> Alg n b

Transfer to Gross Profit. Consolidates Sales, Purchases, WageExpenditure, Depreciation, and ValueAdded into GrossProfit.

Complexity: O(s) (s = total number of scalar entries)

ordinaryProfitTransfer :: (HatVal n, ExBaseClass b) => Alg n b -> Alg n b

Ordinary Profit Transfer

>>> type Test = Alg Double (HatBase (CountUnit, AccountTitles))
>>> x = 2279.0:@Not:<(Yen,Depreciation) .+ 500475.0:@Not:<(Yen,InterestEarned) :: Test
>>> ordinaryProfitTransfer x
2279.00:@Hat:<(Yen,OrdinaryProfit) .+ 500475.00:@Not:<(Yen,OrdinaryProfit)

retainedEarningTransfer :: (HatVal n, ExBaseClass b) => Alg n b -> Alg n b

Transfer to Retained Earnings. Transfers OrdinaryProfit to RetainedEarnings.

Complexity: O(s) (s = total number of scalar entries)

finalStockTransferStep :: (HatVal n, ExBaseClass b) => Alg n b -> Alg n b

Internal step of the final stock transfer from income statement to retained earnings. Cost accounts are transferred to RetainedEarnings with Hat/Not flipped; revenue accounts are transferred to RetainedEarnings as-is.

Complexity: O(s) (s = total number of scalar entries)

finalStockTransfer :: (HatVal n, ExBaseClass b) => Alg n b -> Alg n b

Final Stock Transfer (closing entries). Transfers all cost and revenue accounts to RetainedEarnings and cancels via the bar operation.

Complexity: O(s) (s = total number of scalar entries)

Orphan instances

HatVal n => Show (n -> n)

Instance details

Methods

showsPrec :: Int -> (n -> n) -> ShowS

show :: (n -> n) -> String

showList :: [n -> n] -> ShowS