ExchangeAlgebra.Journal.Transfer

Description

Released under the OWL license

Package for Exchange Algebra defined by Hiroshi Deguchi.

Exchange Algebra is an algebraic description of bookkeeping system. Details are below.

https://www.springer.com/gp/book/9784431209850

https://repository.kulib.kyoto-u.ac.jp/dspace/bitstream/2433/82987/1/0809-7.pdf

_Note_ : The current version 0.1.0.0 will be completely changed shortly, especially in the accounts settings section.

Documentation

data TransTable n b where

Transfer transformation table

Constructors

NullTable :: (HatVal n, HatBaseClass b) => TransTable n b

TransTable

Fields

:: (HatVal n, HatBaseClass b)

=> { _size :: Size

, _before :: b

Base before transformation

, _transFunc :: n -> n

Value transformation function

, _after :: b

Base after transformation

, _left :: TransTable n b

, _right :: TransTable n b

} -> TransTable n b

isNullTable :: TransTable n b -> Bool

Tests whether the transfer table is empty.

Complexity: O(1)

table :: (HatVal n, HatBaseClass b) => [(b, b, n -> n)] -> TransTable n b

make TransTable from list same as fromList >>> table $ Hat::- Hat:|% (id :: NN.Double - NN.Double) ++ Hat::- Hat:<(Cash) |% id [(Hat:Cash,Hat:<Building,<function),(Hat:Building,Hat:<Cash,<function)]

data TransTableParts b

A part of a transfer rule. Represents a source-to-target base pair in from :-> to form.

(.->) :: HatBaseClass b => b -> b -> TransTableParts b infixr 8

Transfer rule construction operator. from .-> to produces a TransTableParts.

Complexity: O(1)

(|%) :: (HatVal n, HatBaseClass b) => TransTableParts b -> (n -> n) -> [(b, b, n -> n)] infixr 7

Syntax to make list for makeList

>>> Hat:<(Yen,Cash):-> Hat:<(Yen,Building) |% (id :: NN.Double -> NN.Double) ++ Not:<(Yen,Building)  :-> Not:<(Yen, Cash)  |% id
[(Hat:<(Yen,Cash),Hat:<(Yen,Building),<function>),(Not:<(Yen,Building),Not:<(Yen,Cash),<function>)]

transfer :: (HatVal v, HatBaseClass b, Note n) => Journal n v b -> TransTable v b -> Journal n v b

Apply transfer transformations to each Note entry in a Journal. Wildcard portions within tuples are not transformed and retain their original values.

Complexity: O(j * s) (j = number of Notes, s = number of scalar entries per Note)

incomeSummaryAccount :: (Note n, HatVal v, ExBaseClass b) => Journal n v b -> Journal n v b

Compute net income for the current period (Income Summary Account). Calculate the debit-credit difference and add it as NetIncome or NetLoss to the plank Note.

Complexity: O(s) (s = total number of scalar entries)

netIncomeTransfer :: (Note n, HatVal v, ExBaseClass b) => Journal n v b -> Journal n v b

Net income transfer (Journal version). Transfer NetIncome/NetLoss to RetainedEarnings for each Note.

Complexity: O(j * s) (j = number of Notes, s = number of scalar entries per Note)

grossProfitTransfer :: (Note n, HatVal v, ExBaseClass b) => Journal n v b -> Journal n v b

Gross profit transfer (Journal version). Aggregate sales and cost accounts into GrossProfit for each Note.

Complexity: O(j * s)

ordinaryProfitTransfer :: (Note n, HatVal v, ExBaseClass b) => Journal n v b -> Journal n v b

Ordinary Profit Transfer

>>> type Test = Journal String Double (HatBase (CountUnit, AccountTitles))
>>> x = 2279.0:@Not:<(Yen,Depreciation) .| "A" :: Test
>>> y = 500475.0:@Not:<(Yen,InterestEarned) .| "B" :: Test
>>> ExchangeAlgebra.Journal.Transfer.ordinaryProfitTransfer ( x .+ y)
2279.00:@Hat:<(Yen,OrdinaryProfit).|"A" .+ 500475.00:@Not:<(Yen,OrdinaryProfit).|"B"

retainedEarningTransfer :: (Note n, HatVal v, ExBaseClass b) => Journal n v b -> Journal n v b

Retained earnings transfer (Journal version). Transfer OrdinaryProfit to RetainedEarnings for each Note.

Complexity: O(j * s)

finalStockTransfer :: (Note n, HatVal v, ExBaseClass b) => Journal n v b -> Journal n v b

Income summary account (Journal version). Transfer all cost and revenue accounts to RetainedEarnings, then offset using the Bar operation.

Complexity: O(j * s)